requirements.txt
asgiref==3.9.2
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.2
django-formtools==2.5.1
djangorestframework==3.16.1
fonttools==4.60.1
gunicorn==23.0.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.11.0
